Update Asgardia.Constitution.Proposal.txt
I was able to read through and update formatting and grammar on all sections again, and I included all changes up to your most recent one (actually, I just checked the site, and saw that you submitted a change for Art. 37; this does not include that).

I think I'm selecting the correct option here by choosing a new branch so that you can review and push it. I'm really liking where this is going so far. Once all of the grammar and formatting is out of the way, I am going to start focusing on content as well.