Update Asgardia.Constitution.Proposal.txt
Main changes were removal of the hanging spaces at the end of some of the lines. Also reworded a few of the lists to make more sense.

My key concern is the removal of "Earth-" modifiers on time durations. We need a better way to indicate time simply because of the concept on this nation. If we start to have habitations on Mars, for instance, someone who is 30 in Earth years is only about 15 in Martian years. It could drastically change the concept of some of the regulations we've put in place in the constitution. I would think, as our nation began from Earth, it makes sense to standardize the time on Earth until we can come up with a Universal concept that works everywhere.